# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/brands" : {
      "get" : {
        "tags" : [ "restaurant-onboarding-controller" ],
        "operationId" : "getBrands",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListBrandSummaryDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "restaurant-onboarding-controller" ],
        "operationId" : "onboardBrand",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrandOnboardRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBrand"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BrandOnboardRequest" : {
        "required" : [ "bankAccountNumber", "gstin", "ifscCode", "name", "organisationId", "pan" ],
        "type" : "object",
        "properties" : {
          "organisationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "gstin" : {
            "type" : "string"
          },
          "pan" : {
            "type" : "string"
          },
          "cin" : {
            "type" : "string"
          },
          "bankAccountNumber" : {
            "type" : "string"
          },
          "ifscCode" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseBrand" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/Brand"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "Brand" : {
        "required" : [ "id", "name", "organisationId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "organisationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "gstin" : {
            "type" : "string"
          },
          "pan" : {
            "type" : "string"
          },
          "cin" : {
            "type" : "string"
          },
          "bankAccountNumber" : {
            "type" : "string"
          },
          "bankIfsc" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "legalEntityName" : {
            "type" : "string"
          },
          "kycStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "VERIFIED", "REJECTED", "MANUAL_REVIEW", "FAILED" ]
          },
          "bankBeneficiaryName" : {
            "type" : "string"
          },
          "pennyDropStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "VERIFIED", "REJECTED", "MANUAL_REVIEW", "FAILED" ]
          },
          "isGstinVerified" : {
            "type" : "boolean"
          },
          "isBankVerified" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ApiResponseListBrandSummaryDto" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BrandSummaryDto"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "BrandSummaryDto" : {
        "required" : [ "id", "name", "organisationId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "organisationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "kycStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "VERIFIED", "REJECTED", "MANUAL_REVIEW", "FAILED" ]
          },
          "pennyDropStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "VERIFIED", "REJECTED", "MANUAL_REVIEW", "FAILED" ]
          },
          "isGstinVerified" : {
            "type" : "boolean"
          },
          "isBankVerified" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    }
  }
}