# src/main/java/com/fooddelivery/restaurant/controller/RestaurantOnboardingController.java
package com.fooddelivery.restaurant.controller;

import com.fooddelivery.restaurant.dto.*;

import com.fooddelivery.common.dto.ApiResponse;
import com.fooddelivery.restaurant.entity.Brand;
import com.fooddelivery.restaurant.entity.Outlet;
import com.fooddelivery.restaurant.service.RestaurantOnboardingService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import java.util.UUID;
import java.util.Map;
import java.util.HashMap;
import java.time.LocalTime;
import java.util.List;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import org.springframework.security.access.prepost.PreAuthorize;

@RestController
@lombok.extern.slf4j.Slf4j
@lombok.RequiredArgsConstructor
public class RestaurantOnboardingController {
private final RestaurantOnboardingService onboardingService;
    private final com.fooddelivery.common.client.GovernmentIdServiceClient governmentIdClient;


    private final com.fooddelivery.common.service.RateLimitingService rateLimitingService;

    // Phase 1: Brand Onboarding
    @PostMapping("/api/v1/brands")
    @PreAuthorize("hasRole('RESTAURANT') and @restaurantAccess.onOrganisation(#request.organisationId, authentication, T(com.fooddelivery.common.enums.OrganisationPermission).BUSINESS_APPLY)")
    public ResponseEntity<ApiResponse<Brand>> onboardBrand(java.security.Principal principal, @Valid @RequestBody BrandOnboardRequest request) {
        io.github.bucket4j.Bucket bucket = rateLimitingService.resolveBucket("onboarding_brand:" + (principal != null ? principal.getName() : "anonymous"), 5, 5, java.time.Duration.ofHours(1));
        if (!bucket.tryConsume(1)) {
            return ResponseEntity.status(org.springframework.http.HttpStatus.TOO_MANY_REQUESTS).build();
        }
        Brand brand = onboardingService.onboardBrand(request.getOrganisationId(), request.getName(), request.getGstin(), request.getPan(), request.getCin(), request.getBankAccountNumber(), request.getIfscCode(), request.getLogoUrl());
        // KYC is triggered async via Outbox/Kafka in the onboardingService
        return ResponseEntity.ok(ApiResponse.success(brand, "Brand onboarded successfully. KYC pending."));
    }



    @GetMapping("/api/v1/brands")
    @PreAuthorize("hasRole('RESTAURANT')")
    public ResponseEntity<ApiResponse<List<BrandSummaryDto>>> getBrands(java.security.Principal principal) {
        List<BrandSummaryDto> brands = onboardingService.getBrands(UUID.fromString(principal.getName()))
                .stream().map(BrandSummaryDto::from).toList();
        return ResponseEntity.ok(ApiResponse.success(brands, "Brands retrieved successfully"));
    }



}